requirements.txt
Authlib==0.15.5
connexion==2.10.0
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
psycopg2==2.9.1
PyJWT==2.3.0
PyNaCl==1.3.0
SQLAlchemy-Utils==0.38.2
swagger-ui-bundle==0.0.9
